% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{Activation}
\alias{Activation}
\title{Activation functions enum}
\format{An object of class \code{list} of length 5.}
\usage{
Activation
}
\description{
Available activation functions:
\itemize{
\item \code{Activation$IDENTITY}
\item \code{Activation$SIGMOID}
\item \code{Activation$TANH}
\item \code{Activation$SOFTMAX}
\item \code{Activation$RELU}
}
}
\keyword{datasets}
