% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{Cost}
\alias{Cost}
\title{Cost functions enum}
\format{An object of class \code{list} of length 3.}
\usage{
Cost
}
\description{
Available cost functions:
\itemize{
\item \code{Cost$QUADRATIC}
\item \code{Cost$BINARY_CROSS_ENTROPY}
\item \code{Cost$CATEGORICAL_CROSS_ENTROPY}
}
}
\keyword{datasets}
