% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolores.R
\name{Dolores}
\alias{Dolores}
\title{Dolores R6 Class}
\description{
Dolores R6 Class

Dolores R6 Class
}
\details{
Dolores is a simple feedforward neural network implementation.
It uses backpropagation as learning mechanism.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{config}}{List of configuration values}

\item{\code{weights}}{List of weights matrices}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Dolores$new()}}
\item \href{#method-train}{\code{Dolores$train()}}
\item \href{#method-validate}{\code{Dolores$validate()}}
\item \href{#method-calculate}{\code{Dolores$calculate()}}
\item \href{#method-clone}{\code{Dolores$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new Dolores object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dolores$new(
  layers,
  learning_rate = 0.01,
  batch_size = 1,
  epochs = 10,
  cost = Cost$QUADRATIC
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layers}}{List of layers configurations created by \code{\link[=layer]{layer()}}}

\item{\code{learning_rate}}{Learning rate. Defaults to \code{.01}}

\item{\code{batch_size}}{Batch size. Defaults to \code{1}}

\item{\code{epochs}}{Number of epochs. Defaults to \code{10}}

\item{\code{cost}}{Cost function vector defined as \verb{[function, function_derivative]}

Check out \code{\link{Cost}} enum for available functions. Defaults to \code{Cost$QUADRATIC}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Dolores} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\subsection{Method \code{train()}}{
Train neural network.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dolores$train(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Training data in matrix or data frame format.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Current \code{Dolores} instance.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\subsection{Method \code{validate()}}{
Validate neural network.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dolores$validate(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Test data in matrix or data frame format.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns output of cost function. It's usually a list of total and mean errors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate"></a>}}
\subsection{Method \code{calculate()}}{
Feed data into network and calculate output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dolores$calculate(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Input data in matrix or data frame format.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Calculated output values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dolores$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
